+++ test/unit/fuzz-tools.test.js
@@ -378,6 +378,7 @@ import * as carpentierMrV398 from '../../lib/carpentier-mr-v398.js';
 import * as bismuthCorletteV399 from '../../lib/bismuth-corlette-v399.js';
 import * as nyhusHerniaV400 from '../../lib/nyhus-hernia-v400.js';
 import * as zargarCausticV401 from '../../lib/zargar-caustic-v401.js';
+import * as laugeHansenV402 from '../../lib/lauge-hansen-v402.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -701,6 +702,7 @@ const MODULES = {
   'bismuth-corlette-v399.js': bismuthCorletteV399,
   'nyhus-hernia-v400.js': nyhusHerniaV400,
   'zargar-caustic-v401.js': zargarCausticV401,
+  'lauge-hansen-v402.js': laugeHansenV402,
   'tb-testing.js': tbTesting,
 };